requirement.txt
faiss-cpu==1.11.0.post1
google-generativeai==0.8.5
numpy==2.1.3
python-docx==1.2.0
PyMuPDF==1.26.3